Clean up - debugging.
Minor changes for debugging types.

+++ chip8emu/chip8emu/Stack.cpp
@@ -13,7 +13,7 @@ Stack::~Stack()
 Stack::Stack(int size)
 {
 	this->size = size;
-	values = new int[size];
+	values = new unsigned short[size];
 	memset(values, 0, size);
 	top = -1;
 }
@@ -30,7 +30,7 @@ bool Stack::isFull()
 	}
 }
 
-void Stack::push(int x)
+void Stack::push(unsigned short x)
 {
 	if (!isFull())
 	{
@@ -51,9 +51,9 @@ bool Stack::isEmpty()
 	}
 }
 
-int Stack::pop()
+unsigned short Stack::pop()
 {
-	int retVal = 0;
+	unsigned short retVal = 0;
 	if (!isEmpty())
 	{
 		retVal = values[top];
